+++ backend/canisters/group/api/src/queries/api_key.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
-use types::{ApiKey, UserId};
+use types::UserId;
 
 #[ts_export(group, api_key)]
 #[derive(Serialize, Deserialize, Debug)]
@@ -9,10 +9,19 @@ pub struct Args {
 }
 
 #[ts_export(group, api_key)]
-#[allow(clippy::large_enum_variant)]
 #[derive(Serialize, Deserialize, Debug)]
 pub enum Response {
-    Success(ApiKey),
+    Success(String),
     NotAuthorized,
     NotFound,
+}
+
+impl From<Response> for types::c2c_bot_api_key::Response {
+    fn from(value: Response) -> Self {
+        match value {
+            Response::Success(s) => types::c2c_bot_api_key::Response::Success(s),
+            Response::NotAuthorized => types::c2c_bot_api_key::Response::NotAuthorized,
+            Response::NotFound => types::c2c_bot_api_key::Response::NotFound,
+        }
+    }
 }