+++ backend/libraries/types/src/c2c_bot_api_key.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Debug)]
+pub enum Response {
+    Success(String),
+    NotAuthorized,
+    NotFound,
+}